requirements.txt
baidu-aip
chardet
opencv-python
pillow